Ensure we get an error if the 'Position' JS changes
Matching some in-page JavaScript with a regular expression is pretty
brittle, so I'd like to make sure that if we get any unexpected value in
that field, then there's an exception that causes the scraper to fail so
that a developer knows to investigate. This will catch the JS being
changed so the regular expression no longer matches or new values other
than "Hon" or "Sen" in the JS conditional.