Extract a JS string which is used for the displayed 'Position'
The "Position" cell in the table is populated based on a literal string
in the JavaScript if it would otherwise be empty; this commit extracts
that string and outputs it as a 'js_position' field.

Matching some in-page JavaScript with a regular expression is pretty
brittle, so I'd like to make sure that if we get any unexpected value in
that field, then there's an exception that causes the scraper to fail so
that a developer knows to investigate. This will catch the JS being
changed so the regular expression no longer matches or new values other
than "Hon" or "Sen" in the JS conditional.