Scrape members of the Senate, not just House of Representatives
This commit changes the scraper to also include members of the Senate.
This means that the instructions.json for the House of Representatives
of Nigeria in everypolitician-data would need to be updated to only
SELECT those with a title of 'Hon.' in order to preserve the
original behaviour; however, @tmtmtmtm has pointed out, SELECTing
on the 'Position' in the page is likely more reliable, which would
suggest using the new 'js_position' column.