Remove the possibility of scraping a member's URL multiple times
Previously, it would have been the case that if the Nigerian National
Assembly website was fixed and one of the people we added manually to
the list of URLs to be scraped reappeared in the search index, their
page would end up being scraped twice.

This commit uses @tmtmtmtm's suggestion to only collect URLs from the
search page results (since the name was unused anyway) and use the |
operator on the array of URLs to ensure we don't add URLs that are
already present.