Extract the fullest version of the person's name from the page
The name in the info box (which is the 'name' field) is a shorter
version of the person's name than that used in the <h1> on the page. It
would be good to have the scraper extract the fullest version of their
name too, which is often helpful for disambiguation.