Add a special case end_date for Philip Shaibu, who is now a governor
The ideal way of dealing with a case like this, where:

 (a) A person has stepped down from their post midway through a term
 (b) But they still have a page on the website
 (c) And there is no end date scrapeable from that page

... would be to add the end date to Wikidata. However, at the moment
there's no way to say in the instructions.json in everypolitician-data
that Wikidata information for a membership can override or update a
membership from a previous source. Until there's time to work on that
solution, we're just hardcoding the end_date of this representative in
the scraper.