Switch back to non-secure URLs
There's currently a problem with the HTTPS version of the URLs because
the website is serving up an incomplete certificate chain, so openssl
refuses to connect to it. This means that we can't scrape it properly
from Ruby that's compiled against openssl (which is pretty standard).

Interestingly it does work under certain conditions, such as viewing the
site in the browser or using a curl compiled with LibreSSL (such as the
one that's shipped with MacOS), but not a curl that's compiled with
openssl.

I have sent a message using the websites feedback form to make them
aware of this issue, so hopefully it will be resolved, but in the
meantime the site does still work with the non-secure version of the
URLs, so lets use that so we can develop against it locally.